requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
h11==0.14.0
idna==3.6
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.9.0
fsspec==2023.4.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
pillow==10.2.0
requests==2.31.0
sympy==1.12
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
torchvision==0.17.0+cpu
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.2.0
fastapi==0.109.2
filelock==3.9.0
fonttools==4.48.1
fsspec==2023.4.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
pycocotools==2.0.7
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.7
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
